+++ src/main/java/fyi/sorenneedscoffee/derg_zero/boosters/data/models/QueuedBooster.java
@@ -2,13 +2,16 @@ package fyi.sorenneedscoffee.derg_zero.boosters.data.models;
 
 import java.time.temporal.ChronoUnit;
 
+/**
+ * @author SorenNeedsCoffee (github.com/sorenneedscoffee)
+ */
 public class QueuedBooster {
     public final int id;
-    public final double multiplier;
+    public final float multiplier;
     public final long duration;
     public final ChronoUnit unit;
 
-    public QueuedBooster(int id, double multiplier, long duration, ChronoUnit unit) {
+    public QueuedBooster(int id, float multiplier, long duration, ChronoUnit unit) {
         this.id = id;
         this.multiplier = multiplier;
         this.duration = duration;
